backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
httpx
supabase
python-dotenv